Replace print by the Python logging framework
We use the logger "fishtest" as configured in production.ini.

For messages to the event log we use the child logger
"fishtest.event", also configured in "production.ini".

Log messages are formatted as follows:

INFO [fishtest] [waitress-0:rundb.py:failed_task:1446] <...message...>

Also: move scheduler to a separate file scheduler.py.
