+++ server/development.ini
@@ -35,10 +35,10 @@ port = 6542
 ###
 
 [loggers]
-keys = root, fishtest
+keys = root, fishtest, fishtest.event
 
 [handlers]
-keys = console
+keys = console, events
 
 [formatters]
 keys = generic
@@ -52,11 +52,20 @@ level = DEBUG
 handlers =
 qualname = fishtest
 
+[logger_fishtest.event]
+level = DEBUG
+handlers = events
+qualname = fishtest.event
+
 [handler_console]
 class = StreamHandler
 args = (sys.stderr,)
 level = NOTSET
 formatter = generic
 
+[handler_events]
+class = fishtest.utils.EventHandler
+formatter = generic
+
 [formatter_generic]
-format = %(asctime)s %(levelname)-5.5s [%(name)s][%(threadName)s] %(message)s
+format = %(levelname)s [%(name)s] [%(threadName)s:%(filename)s:%(funcName)s:%(lineno)s]  %(message)s
