+++ server/fishtest/views.py
@@ -14,6 +14,7 @@ import requests
 from fishtest.schemas import RUN_VERSION, runs_schema, short_worker_name
 from fishtest.util import (
     email_valid,
+    event_log,
     extract_repo_from_link,
     format_bounds,
     format_date,
@@ -22,6 +23,7 @@ from fishtest.util import (
     get_tc_ratio,
     github_repo_valid,
     is_sprt_ltc_data,
+    logger,
     password_strength,
     update_residuals,
 )
@@ -290,7 +292,7 @@ def upload(request):
         )
         return {}
     except Exception as e:
-        print("Error reading the network file:", e)
+        logger.error(f"Error reading the network file: {str(e)}")
         request.session.flash("Error reading the network file", "error")
         return {}
     if request.rundb.get_nn(filename):
@@ -314,7 +316,7 @@ def upload(request):
         with open(os.path.expanduser("~/fishtest.upload"), "r") as f:
             upload_server = f.read().strip()
     except Exception as e:
-        print("Network upload not configured:", e)
+        logger.error(f"Network upload not configured: {str(e)}")
         request.session.flash("Network upload not configured", "error")
         return {}
     try:
@@ -323,10 +325,10 @@ def upload(request):
         response = requests.post(upload_server, files=files, timeout=HTTP_TIMEOUT * 20)
         response.raise_for_status()
     except ConnectionError as e:
-        print("Failed to connect to the net server:", e)
+        logger.error(f"Failed to connect to the net server: {str(e)}")
         error = "Failed to connect to the net server"
     except HTTPError as e:
-        print("Network upload failed:", e)
+        logger.error("Network upload failed: {str(e)}")
         if response.status_code == 409:
             error = "Post request failed: network {} already uploaded".format(filename)
         elif response.status_code == 500:
@@ -336,7 +338,7 @@ def upload(request):
         else:
             error = "Post request failed: other HTTP error"
     except Exception as e:
-        print("Error during connection:", e)
+        logger.error(f"Error during connection: {str(e)}")
         error = "Post request for the network upload failed"
 
     if error:
@@ -421,7 +423,7 @@ def signup(request):
             ).json()
             if "success" not in response or not response["success"]:
                 if "error-codes" in response:
-                    print(response["error-codes"])
+                    logger.error(response["error-codes"])
                 request.session.flash("Captcha failed", "error")
                 return {}
 
@@ -730,7 +732,7 @@ def get_master_info(url):
         response = requests.get(url)
         response.raise_for_status()
     except Exception as e:
-        print(f"Exception getting commits:\n{e}")
+        logger.error(f"Exception getting commits:\n{str(e)}")
         return None
 
     bench_search = re.compile(r"(^|\s)[Bb]ench[ :]+([1-9]\d{5,7})(?!\d)")
@@ -1407,15 +1409,11 @@ def tests_delete(request):
         try:
             validate(runs_schema, run, "run")
         except ValidationError as e:
-            message = (
-                f"The run object {request.POST['run-id']} does not validate: {str(e)}"
-            )
-            print(message, flush=True)
+            message = f"The run object {run_id} does not validate: {str(e)}"
             if "version" in run and run["version"] >= RUN_VERSION:
-                request.actiondb.log_message(
-                    username="fishtest.system",
-                    message=message,
-                )
+                event_log.error(message)
+            else:
+                logger.info(message)
         request.rundb.buffer(run, True)
 
         request.actiondb.delete_run(
@@ -1767,7 +1765,7 @@ def tests(request):
             try:
                 last_tests = homepage_results(request)
             except Exception as e:
-                print("Overview exception: " + str(e))
+                logger.error(f"Overview exception: {str(e)}")
                 if not last_tests:
                     raise e
             finally:
