+++ server/production.ini
@@ -40,10 +40,10 @@ threads = 4
 ###
 
 [loggers]
-keys = root, fishtest
+keys = root, fishtest, fishtest.event
 
 [handlers]
-keys = console
+keys = console, events
 
 [formatters]
 keys = generic
@@ -53,9 +53,14 @@ level = ERROR
 handlers = console
 
 [logger_fishtest]
-level = WARN
+level = DEBUG
 handlers =
 qualname = fishtest
+
+[logger_fishtest.event]
+level = DEBUG
+handlers = events
+qualname = fishtest.event
 
 [handler_console]
 class = StreamHandler
@@ -63,5 +68,10 @@ args = (sys.stderr,)
 level = NOTSET
 formatter = generic
 
+[handler_events]
+class = fishtest.util.EventHandler
+level = NOTSET
+formatter = generic
+
 [formatter_generic]
-format = %(asctime)s %(levelname)-5.5s [%(name)s][%(threadName)s] %(message)s
+format = %(levelname)s [%(name)s] [%(threadName)s:%(filename)s:%(funcName)s:%(lineno)s]  %(message)s
